refactor code as per comments
