+++ litellm/utils.py
@@ -7216,7 +7216,7 @@ class ProviderConfigManager:
         elif litellm.LlmProviders.COMETAPI == provider:
             return litellm.CometAPIEmbeddingConfig()
         elif litellm.LlmProviders.SAGEMAKER == provider:
-            from litellm.llms.sagemaker.completion.transformation import SagemakerEmbeddingConfig
+            from litellm.llms.sagemaker.embedding.transformation import SagemakerEmbeddingConfig
             return SagemakerEmbeddingConfig.get_model_config(model)
         return None
